pyproject.toml
[project]
name = "vectordb"
version = "0.1.0"
description = ""
authors = [
    {name = "nagendrasisodiya",email = "3n3singh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "sentence-transformers (==4.1.0)",
    "chromadb (>=1.4.0,<2.0.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "ibm-watsonx-ai (>=1.4.11,<2.0.0)",
    "load-dotenv (==0.1.0)",
    "langchain (>=1.2.0,<2.0.0)",
    "langchain-ibm (>=1.0.2,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "pypdf (>=6.5.0,<7.0.0)",
    "lark (>=1.3.1,<2.0.0)",
    "posthog (>=2.4.0,<6.0.0)",
    "llama-index (>=0.14.10,<0.15.0)",
    "llama-index-embeddings-huggingface (>=0.6.1,<0.7.0)",
    "llama-index-llms-ibm (>=0.6.1,<0.7.0)",
    "llama-index-retrievers-bm25 (>=0.6.5,<0.7.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "stemmer (>=0.0.4,<0.0.5)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "tensorflow-hub (>=0.16.1,<0.17.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

